refactor (db options): Simplify the generation of dbw Options
